get existing tower secrets
