+++ lib/client/queries.js
@@ -392,22 +392,3 @@ export const HCMNamespace = gql`
     }
   }
 `
-export const HCMSecretList = gql`
-  query applicationSecrets {
-    items: applicationSecrets {
-      metadata {
-        name
-      }
-    }
-  }
-`
-
-export const HCMSecret = gql`
-  query applicationSecrets($namespace: String!) {
-    items: applicationSecrets(namespace: $namespace) {
-      metadata {
-        name
-      }
-    }
-  }
-`
