+++ nls/platform.properties
@@ -108,6 +108,9 @@ description.ansible.job.status=Ansible Tower Job status
 description.ansible.job.status.empty= Ansible Tower job was not executed.
 description.ansible.job.status.empty.err= Ansible task was not executed. Check the Subscription YAML for status errors.
 description.ansible.job.url=Ansible Tower Job URL
+creation.app.section.prePost=Set pre and post deployment tasks
+creation.app.loading.secrets=Loading secrets
+creation.app.loading.secrets.ns.err=To select existing secrets, application Namespace must be set first
 
 description.ansible.task.status=Ansible Task status
