+++ tests/jest/config/platform-properties.json
@@ -96,6 +96,9 @@
   "description.ansible.job.status.empty": "Ansible Tower job was not executed.",
   "description.ansible.job.status.empty.err": "Ansible task was not executed. Check the Subscription YAML for status errors.",
   "description.ansible.job.url": "Ansible Tower Job URL",
+  "creation.app.section.prePost": "Set pre and post deployment tasks",
+  "creation.app.loading.secrets": "Loading secrets",
+  "creation.app.loading.secrets.ns.err": "To select existing secrets, application Namespace must be set first",
   "description.ansible.task.status": "Ansible Task status",
   "description.noSubs.title": "No subscriptions found",
   "description.noChannels.title": "No channels found",
@@ -256,8 +259,8 @@
   "tooltip.creation.app.helmrepo.package.version": "The version or versions for the deployable. You can use a range of versions in the form >1.0, or <3.0.",
   "tooltip.creation.app.ansibleSecretName": "If using Ansible pre or post tasks, enter the Ansible Tower secret name to be used by these tasks. Create new secret or select from the existing secrets defined in the application namespace. ",
   "creation.app.ansible.secret.name": "Ansible Tower secret name",
-  "creation.app.ansible.secret.host": "Enter Ansible Tower host",
-  "creation.app.ansible.secret.token": "Enter Ansible Tower token",
+  "creation.app.ansible.secret.host": "Ansible Tower host",
+  "creation.app.ansible.secret.token": "Ansible Tower token",
   "dashboard.additionalDetails": "View additional details",
   "dashboard.card.clusters": "Cluster health",
   "dashboard.card.deployable": "Deployable",
